cmd/compile: use logopt to track type shapes allocated
This emits a compact descriptor with single characters
for each type size/alignment.  This uses the -json flag,
but with version 1 (which adds these records those already
emitted, about a 10% increase and not necessarily desirable
for other clients of the interface).

1, 2, 4, 8 are corresponding size and align for int and float.
s is string
S is slice
I is interface
c is complex64
C is complex128
P = pointer
Q = pointer to 8-byte-aligned (for compressed pointer experiments)
Structs are enclosed in {}
Arrays are indicated with a leading [number]

Change-Id: I1622ac50dd192ccdd4aefc1777d36d188d398a80