cmd/internal/obj/ppc64: remove C_UCON optab matching class
This class was an odd workaround to the limitations of the matching
rules when assembling opcodes. Instead, use the preprocess pass to
convert opcodes into the IS (immediate shifted) form of a given
opcode.

This allows constant operands in the range of [0x80010000, 0xFFFF0000]
divisible by 0x10000 used be OR/AND/XOR/MOVD/MOVW to be assembled into
one opcode. The C_UCON class was limited to 31 bit unsigned integers.

Note, ADDIS sign-extends the constant argument, all others do not.

For matching opcodes, this means:
	MOVD $is<<16,Rx becomes ADDIS $is,Rx or ORIS $is,Rx
	MOVW $is<<16,Rx becomes ADDIS $is,Rx
	ADD $is<<16,[Rx,]Ry becomes ADDIS $is[Rx,]Ry
	OR $is<<16,[Rx,]Ry becomes ORIS $is[Rx,]Ry
	XOR $is<<16,[Rx,]Ry becomes XORIS $is[Rx,]Ry

Change-Id: I1a988d9f52517a04bb8dc2e41d7caf3d5fff867c