[release-branch.go1.21] go/types, types2: use a non-generic clone helper
Go 1.20 was adopted as the minimum bootstrap toolchain for Go 1.22,
which made it possible for CL 532157 to introduce a generic clone
function in cmd/compile/internal/types2.

On release-branch.go1.21, types2 needs to be capable of being built
with Go 1.17.13 (the minimum bootstrap toolchain for Go 1.21), so
convert the generic clone helper into a non-generic one. Note that
go/types could use generics, but perhaps keeping them in sync is
slightly preferable.

For #63339.

Change-Id: I0e77584a6065f7626bbc189af6447cc96268d26c