cmd/go: replace mkdirCache with os.Stat
We're about to decouple Mkdir from Builder, which means we won't have
a good place to put this cache. This cache was introduced long ago in
CL 5502055 along with many other changes and no particular
justification was given at the time.

Switch to just using Stat to see if the directory already exists.
Given that Stat is also the first thing os.MkdirAll does, the benefit
of this is that it avoids printing lots of unnecessary "mkdir -p"
commands.

For #62067.

Change-Id: I1f80592cbcdfc56c6185046380019b1823dacc0b