crypto/rand,runtime: switch RtlGenRandom for ProcessPrng
RtlGenRandom is a semi-undocumented API, also known as
SystemFunction036, which we use to generate random data on Windows.
It's definition, in cryptbase.dll, is an opaque wrapper for the
documented API ProcessPrng. Instead of using RtlGenRandom, switch to
using ProcessPrng, since the former is simply a wrapper for the latter,
there should be no practical change on the user side, other than a minor
change in the DLLs we load.

This matches the decision taken by BoringSSL: https://github.com/google/boringssl/commit/e79649ba4d515a1b5f57e397431d81f079ff158e

Change-Id: Ie6891bf97b1d47f5368cccbe92f374dba2c2672a