cmd/go: ensure specifying the buildmode does not affect reproducibility
The logic for deciding what buildmode should be used when one was not
specified differed when building and embedding buildinfo. Futhermore,
the actual buildmode was never used in the build ID, only the -buildmode
build flag value. If a user built two binaries with the same source code
but didn't pass a buildmode when building one binary and explicitly
passed the default buildmode for the GOOS and GOARCH for the second
binary, those binaries would not have identical contents.

Fixes 63559.

Change-Id: If7fa592b7790672960faf60c1007c0fed017a77f