cmd/compile: import methods which is exported in another package but not used and make more functions inlined
The public global variable in package b may use the public type in package a. When package c imports package b but does not import package a, and calls method of the public type in package a through the global variable in package b. If the method is calculated to be inlinable, and unified ir is disable, then it will be inlined. However, if unfied ir is enable, it is not inline at this time. This patch is dedicated to fixing this issue.

Fixes #63687

Change-Id: I759ac6a9e5b5016e35f95cbdc84ae30aa91d22c1