cmd/compile/internal/inline: rework use of ir.StaticValue
When running the code to compute function properties that feed
inlining heuristics, the existing heuristics implementation makes
fairly extensive use of ir.StaticValue and ir.Reassigned to sharpen
the analysis. These calls turn out to cause a significant compile time
increase, due to the fact that each call can potentially walk every
node in the IR for the function. To help with this problem, switch the
heuristics code over to using the new "batch mode" reassignment helper
added in the previous CL. In addition, skip computing accurate
'result' or 'param' flags for large functions, since they are very
unlikely to be inline candidates (this also saves some compile time).

Change-Id: Ib15a62416134386e34b7cfa1130a4b413a37b225