net/http: prevent redirect loop in serveFile if "/" is a normal file
When FileServer(Dir("file")) is used where "file" is a normal file and
not a directory, the server enters a redirect loop.

If serveFile is trying to serve "/" from a FileSystem and it happens to
be a normal file then we can avoid the redirect loop to ../ and serve
the file itself. This is the same behavior of newer apis in net/http
like ServeFileFS.

Fixes #63769

Change-Id: I81e289444e7d0bd72189c2e7b763f5540333e2d0
GitHub-Last-Rev: c8c8bf42c3eb7bebce702b58155b966efdc8f43c
GitHub-Pull-Request: golang/go#63860