cmd/compile/internal/ir: add batch-mode reassignment detection oracle
DO NOT SUBMIT

[Still TBD: keep this in the ir package? move to inlheur? add tests?]

Add a new helper ReassignOracle, useful for doing "batch mode"
reassignment analysis, e.g. deciding whether a given ir.Name or (chain
of names) has a single definition an is never reassigned. The intended
usage model is to create and initialize a ReassignOracle for a given
function, then make a series of queries using it (with the
understanding that changing/mutating the func body IR can invalidate
the info cached in the oracle). This oracle is intended to provide the
same sort of analysis that ir.StaticValue and ir.Reassigned carry out,
but at a greatly reduced cost in compile time.

Note: new helper not yet used; will be hooked in as part of a
subsequent CL.

Change-Id: I283e748e440a9f595df495f6aa48ee9c498702d9