mime/multipart, net/textproto: introduce ErrMessageTooLarge for better error handling
This change modifies the handling of errors when a message is too large in
both mime/multipart and net/textproto packages.
Previously, these packages created a new error with the same string each
time a message was deemed too large.
With this change, they now use a common exported error variable.

The purpose of this modification is to allow proper error identification
by comparing it directly using errors.Is(),
instead of matching error strings which can be error-prone and less efficient.

In addition, related TODO comments that were suggesting this type of improvement
have been removed from the codebase as well.

This PR will be imported into Gerrit with the title and first
comment (this text) used to generate the subject and body of
the Gerrit change.

Change-Id: I891a18cb313651c713a4d304d1a7ad655a7c4601
GitHub-Last-Rev: e0af57921dbc7a51657d513c363aa38f4e4542f1
GitHub-Pull-Request: golang/go#63606