cmd/trace: add basic support for v2 traces
This change refactors the cmd/trace package and adds basic support for
v2 traces.

The following features of note are missing in this CL and will be
implemented in follow-up CLs:
- Flow events in the trace viewer
- The focustask filter for the trace viewer
- The taskid filter for the trace viewer
- The goid filter for the trace viewer
- The network lane in the trace viewer
- gcWaiting state for Goroutine Metrics
- Pprof profiles
- The MMU graph
- The goroutine analysis pages
- The task analysis pages
- The region analysis pages

Most of this CL was written and implemented by Felix Geisendörfer
(felix.geisendoerfer@datadoghq.com), I'm just landing it. :)

For #60773.
For #63960.

Change-Id: I3626e22473227c5980134a85f1bb6a845f567b1b